alchemy v0.1.5 Alchemy.Client

Represents a Client connection to the Discord API. This is the main public interface for the library.

Functions

current_guilds()

Specs

current_guilds :: {:ok, [Alchemy.UserGuild.t]}

Get’s a list of guilds the client is currently a part of.

Examples

{:ok, guilds} = Task.await Client.current_guilds

edit_profile(options)

Specs

edit_profile(user_name: String.t, avatar: String.t) :: {:ok, Alchemy.User.t}

Edits the client’s user_name and/or avatar.

Options

user_name - A string specifiying the new user_name for the client

avatar - A link to an image for the client’s avatar

Examples

# Will edit "behind the scenes"
Client.edit_profile(username: "NewGuy", avatar: "imgur.com/image.jpeg")

iex> {:ok, user} = Task.await Client.edit_profile(username: "NewName")
{:ok, Alchemy.Discord.Users.User%{....

get_user(client_id)

Specs

get_user(String.t) :: {:ok, Alchemy.User.t}

Gets a user by their client_id, returns {:ok, %User{}}

"@me" can be passed to get the info relevant to the Client.

Examples

iex> {:ok, user} = Task.await Client.get_user('client_id')
{:ok, Alchemy.Discord.Users.User%{....

leave_guild(guild_id)

Specs

leave_guild(String.t) :: {:ok, :none}

Makes the client leave a guild.

Examples

Client.leave_guild

start(token)

Starts up a new Client with the given token.