alchemy v0.1.5 Alchemy.UserGuild

A shortened version of a Guild struct, through the view of a User.

id

represents a guild’s id

name

represents a guild’s name

icon

a string representing the guild’s icon hash

owner

whether the user linked to the guild owns it

permissions

bitwise of the user’s enabled/disabled permission

Types

t :: %Alchemy.UserGuild{icon: String.t, id: String.t, name: String.t, owner: Boolean, permissions: Integer}